Show logs in an autoscrolling container
