+++ renderer/components/dashboard/__tests__/running.test.js
@@ -3,15 +3,12 @@
  */
 
 import React from 'react'
-import { theme } from 'ooni-components'
-import { ThemeProvider } from 'styled-components'
-import { IntlProvider } from 'react-intl'
-import { render, cleanup, screen, fireEvent } from '@testing-library/react'
-import English from '../../../../lang/en.json'
+
+import { screen, render, cleanup, fireEvent } from '../../../../test/unit/utils'
 import { ipcRenderer } from 'electron'
 import { testGroups } from '../../nettests'
 
-import Running, { Log, ToggleLogButton } from '../running'
+import Running from '../running'
 
 // Mocking useRouter().push()
 const mockRouterPush = jest.fn()
@@ -23,78 +20,8 @@ jest.mock('next/router', () => ({
   },
 }))
 
-const renderComponent = (component, locale = 'en', messages = English) => {
-  return render(
-    <IntlProvider locale={locale} messages={messages}>
-      <ThemeProvider theme={theme}>{component}</ThemeProvider>
-    </IntlProvider>
-  )
-}
-
-describe('Tests for Log component', () => {
-  afterEach(() => {
-    cleanup()
-  })
-
-  test('Log component renders lines correctly', async () => {
-    const onToggleLog = jest.fn()
-    const logOpen = true
-    const logLines = [
-      'Running tests',
-      'iplookup: using ubuntu',
-      'sessionresolver: http3://dns.google/dns-query... <nil>',
-    ]
-    renderComponent(
-      <Log lines={logLines} onToggleLog={onToggleLog} open={logOpen} />
-    )
-    const logLine1 = screen.getByText(/Running tests/)
-    const logLine2 = screen.getByText(/iplookup: using ubuntu/)
-    const logLine3 = screen.getByText(
-      /sessionresolver: http3:\/\/dns.google\/dns-query... <nil>/
-    )
-    expect(logLine1).toBeInTheDocument()
-    expect(logLine2).toBeInTheDocument()
-    expect(logLine3).toBeInTheDocument()
-  })
-})
-
-describe('Tests for ToggleLogButton component', () => {
-  afterEach(() => {
-    cleanup()
-  })
-
-  test('ToggleLogButton shows "Close Log" when is open and triggers callback on click', async () => {
-    const open = true
-    const onToggleLog = jest.fn()
-    renderComponent(<ToggleLogButton onClick={onToggleLog} open={open} />)
-    const closeLogButton = screen.getByText(
-      English['Dashboard.Running.CloseLog']
-    )
-    const showLogButton = screen.queryByText(
-      English['Dashboard.Running.ShowLog']
-    )
-    expect(closeLogButton).toBeInTheDocument()
-    expect(showLogButton).not.toBeInTheDocument()
-    const toggleLog = screen.getByTestId('toggle-log-button')
-    fireEvent.click(toggleLog)
-    expect(onToggleLog).toHaveBeenCalledTimes(1)
-  })
-  
-  test('ToggleLogButton shows "Show Log" when is open and triggers callback on click', async () => {
-    const open = false
-    const onToggleLog = jest.fn()
-    renderComponent(<ToggleLogButton onClick={onToggleLog} open={open} />)
-    const showLogButton = screen.getByText(English['Dashboard.Running.ShowLog'])
-    const closeLogButton = screen.queryByText(
-      English['Dashboard.Running.CloseLog']
-    )
-    expect(showLogButton).toBeInTheDocument()
-    expect(closeLogButton).not.toBeInTheDocument()
-    const toggeLog = screen.getByTestId('toggle-log-button')
-    fireEvent.click(toggeLog)
-    expect(onToggleLog).toHaveBeenCalledTimes(1)
-  })
-})
+// scrollIntoView is not implemented in jsdom
+window.HTMLElement.prototype.scrollIntoView = jest.fn()
 
 describe('Animation tests for "Running" component', () => {
   afterEach(() => {
@@ -105,7 +32,7 @@ describe('Animation tests for "Running" component', () => {
   testGroupNames.forEach(testGroupName => {
     test(`${testGroupName} test animations load correctly`, async () => {
       const animationData = testGroups[testGroupName].animation
-      renderComponent(<Running testGroupToRun={testGroupName} inputFile={null} />)
+      render(<Running testGroupToRun={testGroupName} inputFile={null} />)
       const animationElement = screen.getByTestId(`running-animation-${testGroupName}`)
       expect(animationElement).toBeInTheDocument()
       expect(animationElement.firstChild.nodeName).toBe('svg')
@@ -117,7 +44,7 @@ describe('Animation tests for "Running" component', () => {
   test('Default test animations load correctly for non-existing test', async () => {
     const testGroupName = 'nonexistent'
     const animationData = testGroups['default'].animation
-    renderComponent(<Running testGroupToRun={testGroupName} inputFile={null} />)
+    render(<Running testGroupToRun={testGroupName} inputFile={null} />)
     const animationElement = screen.getByTestId('running-animation-default')
     expect(animationElement).toBeInTheDocument()
     expect(animationElement.firstChild.nodeName).toBe('svg')
@@ -144,7 +71,7 @@ describe('IPC tests for "ooni" event in "Running" component', () => {
       callback(event, data)
     })
     const testGroupName = 'im'
-    renderComponent(<Running testGroupToRun={testGroupName} inputFile={null} />)
+    render(<Running testGroupToRun={testGroupName} inputFile={null} />)
     const labelTelegram = screen.getByText(/Telegram Test/i)
     const testDetail = screen.getByText(
       'telegram: measure http://95.161.76.100:443/: <nil>'
@@ -165,7 +92,7 @@ describe('IPC tests for "ooni" event in "Running" component', () => {
       callback(event, data)
     })
     const testGroupName = 'im'
-    renderComponent(<Running testGroupToRun={testGroupName} inputFile={null} />)
+    render(<Running testGroupToRun={testGroupName} inputFile={null} />)
     const labelTelegram = screen.getByText(/WhatsApp Test/i)
     const testDetail = screen.getByText(
       'whatsapp: measure tcpconnect://e14.whatsapp.net:443: <nil>'
@@ -184,7 +111,7 @@ describe('IPC tests for "ooni" event in "Running" component', () => {
       callback(event, logObj)
     })
     const testGroupName = 'im'
-    renderComponent(<Running testGroupToRun={testGroupName} inputFile={null} />)
+    render(<Running testGroupToRun={testGroupName} inputFile={null} />)
     const toggeLog = screen.getByTestId('toggle-log-button')
     fireEvent.click(toggeLog)
     const logLine = screen.getByText(logObj.value)
@@ -202,7 +129,7 @@ describe('IPC tests for "ooni" event in "Running" component', () => {
       callback(event, errorObj)
     })
     const testGroupName = 'im'
-    renderComponent(<Running testGroupToRun={testGroupName} inputFile={null} />)
+    render(<Running testGroupToRun={testGroupName} inputFile={null} />)
     const errorLine = screen.getByText(errorObj.message)
     expect(errorLine).toBeInTheDocument()
   })
@@ -221,7 +148,7 @@ describe('IPC tests for other ooniprobe ipc events in "Running" component', () =
         callback(event, testGroupName)
       }
     })
-    renderComponent(<Running testGroupToRun={testGroupName} inputFile={null} />)
+    render(<Running testGroupToRun={testGroupName} inputFile={null} />)
     const toggeLog = screen.getByTestId('toggle-log-button')
     fireEvent.click(toggeLog)
     const logLine = screen.getByText(`Finished running ${testGroupName}`)
@@ -235,7 +162,7 @@ describe('IPC tests for other ooniprobe ipc events in "Running" component', () =
         callback(event)
       }
     })
-    renderComponent(<Running testGroupToRun={testGroupName} inputFile={null} />)
+    render(<Running testGroupToRun={testGroupName} inputFile={null} />)
     expect(mockRouterPush).toHaveBeenCalledWith('/test-results')
   })
 
@@ -247,7 +174,7 @@ describe('IPC tests for other ooniprobe ipc events in "Running" component', () =
         callback(event, errorMessage)
       }
     })
-    renderComponent(<Running testGroupToRun={testGroupName} inputFile={null} />)
+    render(<Running testGroupToRun={testGroupName} inputFile={null} />)
     const toggeLog = screen.getByTestId('toggle-log-button')
     fireEvent.click(toggeLog)
     const logLine = screen.getByText(errorMessage)
